pyproject.toml
[project]
name = "urine-color-analysis"
version = "0.2.0"
description = "Automated urine color analysis system — Raspberry Pi client + Ubuntu server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = []

[project.optional-dependencies]
common = [
    "requests",
    "python-dotenv",
    "pydantic-settings>=2.0",
    "loguru",
    "numpy",
    "pyyaml",
    "Pillow>=10.0",
    "psutil>=5.9",
]
pi = [
    # Hardware packages only available on Raspberry Pi (64-bit ARM Linux).
    # The == 'aarch64' marker prevents uv from resolving these on x86_64 / macOS,
    # which avoids build failures (python-prctl requires libcap-dev headers).
    "picamera2>=0.3.12  ; sys_platform == 'linux' and platform_machine == 'aarch64'",
    "RPi.GPIO>=0.7.1    ; sys_platform == 'linux' and platform_machine == 'aarch64'",
    "raspberrypi-tm1637 ; sys_platform == 'linux' and platform_machine == 'aarch64'",
    "rpi-lcd            ; sys_platform == 'linux' and platform_machine == 'aarch64'",
    "smbus2>=0.4.3      ; sys_platform == 'linux' and platform_machine == 'aarch64'",
]
server = [
    "fastapi",
    "uvicorn[standard]",
    "opencv-python",
    "sqlalchemy",
    "psycopg2-binary",
    "jinja2",
    "python-multipart",
    "ultralytics>=8.4.24",
    "scikit-image>=0.21",
    "scipy>=1.11",
    "torch>=2.0",
    "torchvision>=0.15",
    "segmentation-models-pytorch>=0.3",
    "google-api-python-client>=2.0",
    "google-auth-httplib2>=0.2",
    "google-auth-oauthlib>=1.0",
]

[dependency-groups]
dev = [
    "pytest",
]

[project.scripts]
urine-analyzer = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
